# web/public/openapi.yaml
openapi: 3.0.3
info:
  title: CrabIndex API
  version: 0.1.0 # replaced at runtime with the crabindex crate version
  description: |
    Torrent aggregator API: Jackett-compatible search, native torrent search, Torznab XML,
    statistics and sync protocol.

    **Interactive docs:** [Swagger UI](/swagger)

    **Runtime OpenAPI JSON:** `/swagger/v1/swagger.json` (converted from this YAML file)

    **Static spec (this file):** `/openapi.yaml` (source: `web/public/openapi.yaml`, published to `wwwroot/openapi.yaml`)

    ## Supported tracker slugs

    Parsers / cron (`/cron/{slug}/…`, DevAdmin):
    `anibelka`, `anidub`, `anifilm`, `aniliberty`, `animelayer`, `anistar`, `baibako`,
    `bitru`, `kinozal`, `knaben`, `korsars`, `leproduction`, `lostfilm`, `mazepa`,
    `megapeer`, `nnmclub`, `rudub`, `rutor`, `rutracker`, `selezen`, `subsplease`, `toloka`, `torrentby`,
    `ultradox`, `viruseproject`.

    Web UI icons: `/img/ico/{slug}.ico` (fallback `/img/ico/default.ico`).

    Page-only parsers (`anistar`, `leproduction`, `viruseproject`, `anifilm`) usually do **not**
    appear in `GET /health/background-jobs` (only UpdateTasks/ParseAll-style workers register there).

    ## Authentication

    | Scheme | Usage |
    |--------|--------|
    | `apikey` query | Search, stats (when `apikey` is set in config) |
    | `X-Api-Key` header | Same as query apikey |
    | `Authorization: Bearer` | Same as apikey |
...
    | `devkey` query | Same as `X-Dev-Key` (`?devkey=`) |

...

...

...

...

...

...

...
servers:
...
paths:
...
  /torznab/api:
    get:
      tags: [Torznab]
      summary: Torznab API (caps, search)
      description: |
        Native Torznab XML when `torznab.enable` is true.
        Caps advertise `q`, `imdbid`[, `tvdbid`, `season`, `ep`]. CrabIndex also accepts card fields
        (`title`, `title_original`, `year`, `is_serial`, `genres`), pagination (`limit`/`offset`),
        category aliases (`cat`/`Category`/`categories`/`Category[]`), and tracker filters
        (`Tracker`/`Tracker[]`/`tracker`).
        `t=indexers` returns aggregate `all` plus one indexer entry per `GET /api/v1.0/trackers` name.

        **IMDb / Kinopoisk / TMDB IDs:** `q` as `tt…`/`kp…`/`tmdb…`, a themoviedb.org URL, or `imdbid` →
        Alloha TV API v2 title resolve → FileDB exact search (year ±1 when client `year` absent and `alloha.filterByYear`).
        **TVDB-only** (`tvdbid` / `rid` with no `q`/`imdbid`) returns empty - CrabIndex does not resolve TVDB ids.
        **TV text queries:** `search.stripSeasonEpisode` (default true) searches show name without `S01`/`S01E01`;
        `season`/`ep` still post-filter unless `search.skipSeasonEpisodeFilter` is true.
      operationId: torznabRoot
      parameters:
        - $ref: "#/components/parameters/ApiKeyQuery"
        - $ref: "#/components/parameters/TorznabT"
        - $ref: "#/components/parameters/TorznabQ"
        - $ref: "#/components/parameters/SearchQueryAlt"
        - $ref: "#/components/parameters/TorznabCat"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/CategoryBracket"
        - $ref: "#/components/parameters/Categories"
        - $ref: "#/components/parameters/TorznabImdbId"
        - $ref: "#/components/parameters/ImdbIdAltUnderscore"
        - $ref: "#/components/parameters/ImdbIdAltCamel"
        - $ref: "#/components/parameters/TorznabTvdbId"
        - $ref: "#/components/parameters/TorznabRid"
        - $ref: "#/components/parameters/TorznabSeason"
        - $ref: "#/components/parameters/TorznabEp"
        - $ref: "#/components/parameters/TorznabEpisode"
        - $ref: "#/components/parameters/TorznabTitle"
        - $ref: "#/components/parameters/TorznabTitleOriginal"
        - $ref: "#/components/parameters/Year"
        - $ref: "#/components/parameters/IsSerial"
        - $ref: "#/components/parameters/Genres"
        - $ref: "#/components/parameters/TorznabLimit"
        - $ref: "#/components/parameters/TorznabOffset"
        - $ref: "#/components/parameters/TorznabConfigured"
        - $ref: "#/components/parameters/TrackerFilter"
        - $ref: "#/components/parameters/TrackerArray"
        - $ref: "#/components/parameters/TrackerAlias"
      responses:
        "200":
          description: Torznab XML
          content:
            application/xml:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Torznab disabled
    head:
      tags: [Torznab]
      summary: Torznab HEAD probe
      description: Returns empty Torznab XML body (same content-type as GET).
      operationId: torznabRootHead
      parameters:
        - $ref: "#/components/parameters/ApiKeyQuery"
      responses:
        "200":
          description: Empty XML OK
          content:
            application/xml:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Torznab disabled
  /api/v2.0/indexers/{indexer}/results/torznab/api:
    get:
      tags: [Torznab]
      summary: Torznab API (indexer path)
      description: |
        Jackett-compatible Torznab alias. Path `{indexer}` is normally `all`.
        Jackett's virtual `status:healthy` selector is also treated as aggregate.
        When `{indexer}` is a tracker id (not `all` / `status:healthy`), search results are filtered to that tracker.
        `t=indexers` lists `all` plus per-tracker indexers from the tracker catalog.
        Same query params as `/torznab/api` (caps + CrabIndex card fields, categories, tracker filters),
        including Alloha v2 resolve for `tt…` / `kp…` / `tmdb…` / themoviedb.org URLs / `imdbid`.
      operationId: torznabIndexer
      parameters:
        - name: indexer
          in: path
          required: true
          description: |
            `all` or Jackett `status:healthy` for aggregate search, or a tracker id
            (e.g. `rutracker`) to scope results.
          schema:
            type: string
            example: all
        - $ref: "#/components/parameters/ApiKeyQuery"
        - $ref: "#/components/parameters/TorznabT"
        - $ref: "#/components/parameters/TorznabQ"
        - $ref: "#/components/parameters/SearchQueryAlt"
        - $ref: "#/components/parameters/TorznabCat"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/CategoryBracket"
        - $ref: "#/components/parameters/Categories"
        - $ref: "#/components/parameters/TorznabImdbId"
        - $ref: "#/components/parameters/ImdbIdAltUnderscore"
        - $ref: "#/components/parameters/ImdbIdAltCamel"
        - $ref: "#/components/parameters/TorznabTvdbId"
        - $ref: "#/components/parameters/TorznabRid"
        - $ref: "#/components/parameters/TorznabSeason"
        - $ref: "#/components/parameters/TorznabEp"
        - $ref: "#/components/parameters/TorznabEpisode"
        - $ref: "#/components/parameters/TorznabTitle"
        - $ref: "#/components/parameters/TorznabTitleOriginal"
        - $ref: "#/components/parameters/Year"
        - $ref: "#/components/parameters/IsSerial"
        - $ref: "#/components/parameters/Genres"
        - $ref: "#/components/parameters/TorznabLimit"
        - $ref: "#/components/parameters/TorznabOffset"
        - $ref: "#/components/parameters/TorznabConfigured"
        - $ref: "#/components/parameters/TrackerFilter"
        - $ref: "#/components/parameters/TrackerArray"
        - $ref: "#/components/parameters/TrackerAlias"
      responses:
        "200":
          description: Torznab XML
          content:
            application/xml:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Torznab disabled
    head:
      tags: [Torznab]
      summary: Torznab HEAD probe (indexer path)
      description: Returns empty Torznab XML body (same content-type as GET).
      operationId: torznabIndexerHead
      parameters:
        - name: indexer
          in: path
          required: true
          schema:
            type: string
            example: all
        - $ref: "#/components/parameters/ApiKeyQuery"
      responses:
        "200":
          description: Empty XML OK
          content:
            application/xml:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Torznab disabled
...
components:
  securitySchemes:
    ApiKeyQuery:
      type: apiKey
      in: query
      name: apikey
      description: API key (when configured in init.yaml)
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-Api-Key
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: API key as Bearer token
    DevKeyHeader:
      type: apiKey
      in: header
      name: X-Dev-Key
      description: Admin/dev key - required from internet or tunnel when `devkey` is set; LAN clients may omit it.
    DevKeyQuery:
      type: apiKey
      in: query
      name: devkey
      description: Same as X-Dev-Key (`?devkey=`).
  parameters:
    ApiKeyQuery:
      name: apikey
      in: query
      required: false
      schema:
        type: string
      description: API key (alternative - X-Api-Key header or Bearer)
    TorznabT:
      name: t
      in: query
      schema:
        type: string
        enum: [caps, indexers, search, tvsearch, movie, tv, moviesearch]
      description: Torznab function (`caps`, `indexers`, or search type)
    TorznabQ:
      name: q
      in: query
      schema:
        type: string
      description: |
        Free-text search query. When `search.stripSeasonEpisode` is true (default), inline TV tokens
        (`S01`, `S01E01`, `1x01`, `Season 1`) are stripped and both variants are searched
        (e.g. `silo S01` also searches `silo`). Trailing year is also stripped when
        `search.stripTrailingYear` is true (`укрытие 2023 S01` → `укрытие 2023` → `укрытие`).
        IMDb/KP/TMDB ids (`tt…`/`kp…`/`tmdb…`) are not stripped.
    TorznabCat:
      name: cat
      in: query
      schema:
        type: string
      description: Comma-separated Torznab category ids
    TorznabImdbId:
      name: imdbid
      in: query
      schema:
        type: string
      description: |
        IMDb id (`tt…` or bare digits); aliases `imdb_id`, `imdbId` also accepted.
        Resolved via Alloha TV API v2 to titles before FileDB search.
        Kinopoisk ids are accepted on `q`/`query` as `kp…`; TMDB as `tmdb…` or a themoviedb.org
        `/movie|tv/{id}-…` URL (same Alloha resolve path).
    TorznabTvdbId:
      name: tvdbid
      in: query
      schema:
        type: string
      description: |
        TheTVDB id (tv-search). Advertised in caps; CrabIndex does not resolve TVDB ids.
        A `tvdbid`-only request (no `q`/`imdbid`) returns empty results - send `q` or `imdbid` instead.
    TorznabRid:
      name: rid
      in: query
      schema:
        type: string
      description: Alias for `tvdbid`
    TorznabSeason:
      name: season
      in: query
      schema:
        type: integer
      description: |
        Season number (tv-search post-filter). Skipped when `search.skipSeasonEpisodeFilter` is true
        (clients such as AIOStreams filter episodes themselves).
    TorznabEp:
      name: ep
      in: query
      schema:
        type: integer
      description: |
        Episode number (tv-search post-filter). Skipped when `search.skipSeasonEpisodeFilter` is true.
    TorznabEpisode:
      name: episode
      in: query
      schema:
        type: integer
      description: Alias for `ep`
    TorznabTitle:
      name: title
      in: query
      schema:
        type: string
      description: Localized title (CrabIndex / Lampa card search)
    TorznabTitleOriginal:
      name: title_original
      in: query
      schema:
        type: string
      description: Original title (CrabIndex / Lampa card search)
    TorznabLimit:
      name: limit
      in: query
      schema:
        type: integer
      description: Max results after filters
    TorznabOffset:
      name: offset
      in: query
      schema:
        type: integer
      description: Result offset for pagination
    TorznabConfigured:
      name: configured
      in: query
      schema:
        type: string
      description: |
        Used with `t=indexers`. Omit, empty, or `true` returns the CrabIndex indexer list XML;
        `false` returns an empty `<indexers>` document.
      example: "true"
...
    SearchQueryAlt:
      name: Query
      in: query
      schema:
        type: string
      description: Alternate capitalization of free-text search query (same stripping as `q`)
    ImdbIdAltUnderscore:
      name: imdb_id
      in: query
      schema:
        type: string
      description: Alias for `imdbid`
    ImdbIdAltCamel:
      name: imdbId
      in: query
      schema:
        type: string
      description: Alias for `imdbid`
    Categories:
      name: categories
      in: query
      schema:
        type: array
        items:
          type: integer
      style: form
      explode: true
      description: |
        Torznab/Jackett category ids (repeat or comma-separated).
        Also accepted as `cat`, `Category`, `Category[]`, `Category[n]`.
    Category:
      name: Category
      in: query
      schema:
        type: string
      description: Single or comma-separated category ids (alias for `categories` / `cat`)
    CategoryBracket:
      name: Category[]
      in: query
      schema:
        type: array
        items:
          type: integer
      style: form
      explode: true
      description: Jackett/Lampa category array (e.g. 2000 movies, 5000 TV, 5070 anime)
    TrackerArray:
      name: Tracker[]
      in: query
      schema:
        type: array
        items:
          $ref: "#/components/schemas/TrackerSlug"
      style: form
      explode: true
      description: |
        Jackett-style filter by internal tracker id (e.g. kinozal, rutracker).
        Case-insensitive; matches any comma-separated part of stored trackerName.
    TrackerFilter:
      name: Tracker
      in: query
      schema:
        type: string
        description: Comma-separated `TrackerSlug` values (e.g. `kinozal,korsars`)
      description: |
        Optional tracker filter (comma-separated `TrackerSlug` values); same semantics as `Tracker[]`.
        Also accepted as `tracker` / `Tracker[n]`.
    TrackerAlias:
      name: tracker
      in: query
      schema:
        type: string
        description: Comma-separated `TrackerSlug` values (e.g. `kinozal,rutracker`)
      description: |
        Single-tracker or comma-separated alias for `Tracker[]` / `Tracker`
        (`TrackerSlug` values; case-insensitive; matches any comma-separated part of stored trackerName).
    Genres:
      name: genres
      in: query
      schema:
        type: string
      description: |
        When non-empty, enables CrabIndex card-metadata search mode (same trigger as title/title_original).
        Not used as a genre filter on results.
    Year:
      name: year
      in: query
      schema:
        type: integer
      description: |
        Release year filter / card-search override.
        When omitted on an IMDb/KP/TMDB id query and `alloha.filterByYear` is enabled, Alloha year ±1 is applied.
    IsSerial:
      name: is_serial
      in: query
      schema:
        type: integer
      description: "-1 unknown, 1 movie, 2 serial, 3 tvshow, 4 doc, 5 anime"
  responses:
    Unauthorized:
      description: Missing or invalid API key (when apikey is configured)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    TrackerSlug:
      type: string
      description: Internal tracker id (FileDB / cron / synctrackers)
      enum:
        - anibelka
        - anidub
        - anifilm
        - aniliberty
        - animelayer
        - anistar
        - baibako
        - bitru
        - kinozal
        - knaben
        - korsars
        - leproduction
        - lostfilm
        - mazepa
        - megapeer
        - nnmclub
        - rudub
        - rutor
        - rutracker
        - selezen
        - subsplease
        - toloka
        - torrentby
        - ultradox
        - viruseproject
      example: korsars
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
...
security:
  - ApiKeyQuery: []
  - ApiKeyHeader: []
  - Bearer: []